Force first character of command to be a word character.
This avoids conflicts if your bot has `"*"` as prefix, and other bot
has `"**"` as a prefix. This is rather unlikely to happen on normal
channel, but on such channel as `#botters` (with lots of bots), it's
rather likely.

This wasn't actually implemented in YIBot, but YIBot was safe because
it was ignoring commands it didn't knew. But considering PorygonZ has
command not found handler, it's needed.